+++ providers/implementations/rands/drbg_hmac.c
@@ -18,7 +18,7 @@
 #include "prov/implementations.h"
 #include "prov/provider_ctx.h"
 #include "prov/hmac_drbg.h"
-#include "drbg_local.h"
+#include "prov/drbg.h"
 #include "crypto/evp.h"
 #include "crypto/evp/evp_local.h"
 #include "internal/provider.h"
